# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AKUEB With Hatif API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface User {
  id: number;
  name: string;
  email: string;
  createdAt: string;
}

export interface RegisterBody {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface AdminAuthResponse {
  token: string;
  isAdmin: boolean;
}

export interface Class {
  id: number;
  name: string;
  grade: number;
  /** @nullable */
  description?: string | null;
}

export interface Subject {
  id: number;
  name: string;
  classId: number;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface Material {
  id: number;
  title: string;
  type: string;
  subjectId: number;
  /** @nullable */
  url?: string | null;
  /** @nullable */
  objectPath?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface CreateSubjectBody {
  /** @minLength 1 */
  name: string;
  icon?: string;
  description?: string;
}

export interface UpdateSubjectBody {
  /** @minLength 1 */
  name?: string;
  icon?: string;
  description?: string;
}

export interface CreateMaterialBody {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  type: string;
  description?: string;
  url?: string;
  objectPath?: string;
}

export interface UpdateMaterialBody {
  /** @minLength 1 */
  title?: string;
  type?: string;
  description?: string;
  url?: string;
  objectPath?: string;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}