# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AKUEB With Hatif API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from 'zod';


/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  "status": zod.string()
})


/**
 * @summary Register a new user
 */
export const registerBodyNameMin = 2;

export const registerBodyPasswordMin = 6;



export const RegisterBody = zod.object({
  "name": zod.string().min(registerBodyNameMin),
  "email": zod.string().email(),
  "password": zod.string().min(registerBodyPasswordMin)
})


/**
 * @summary Login
 */
export const LoginBody = zod.object({
  "email": zod.string().email(),
  "password": zod.string()
})

export const LoginResponse = zod.object({
  "user": zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "email": zod.string(),
  "createdAt": zod.string()
}),
  "token": zod.string()
})


/**
 * @summary Logout
 */
export const LogoutResponse = zod.object({
  "message": zod.string()
})


/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "email": zod.string(),
  "createdAt": zod.string()
})


/**
 * @summary List all classes
 */
export const ListClassesResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "grade": zod.number(),
  "description": zod.string().nullish()
})
export const ListClassesResponse = zod.array(ListClassesResponseItem)


/**
 * @summary List subjects for a class
 */
export const ListSubjectsParams = zod.object({
  "classId": zod.coerce.number()
})

export const ListSubjectsResponseItem = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "classId": zod.number(),
  "icon": zod.string().nullish(),
  "description": zod.string().nullish()
})
export const ListSubjectsResponse = zod.array(ListSubjectsResponseItem)


/**
 * @summary List materials for a subject
 */
export const ListMaterialsParams = zod.object({
  "subjectId": zod.coerce.number()
})

export const ListMaterialsResponseItem = zod.object({
  "id": zod.number(),
  "title": zod.string(),
  "type": zod.string(),
  "subjectId": zod.number(),
  "url": zod.string().nullish(),
  "objectPath": zod.string().nullish(),
  "description": zod.string().nullish()
})
export const ListMaterialsResponse = zod.array(ListMaterialsResponseItem)


/**
 * @summary Operator login
 */
export const AdminLoginBody = zod.object({
  "email": zod.string().email(),
  "password": zod.string()
})

export const AdminLoginResponse = zod.object({
  "token": zod.string(),
  "isAdmin": zod.boolean()
})


/**
 * @summary Create a subject (operator only)
 */
export const CreateSubjectParams = zod.object({
  "classId": zod.coerce.number()
})




export const CreateSubjectBody = zod.object({
  "name": zod.string().min(1),
  "icon": zod.string().optional(),
  "description": zod.string().optional()
})


/**
 * @summary Update a subject (operator only)
 */
export const UpdateSubjectParams = zod.object({
  "subjectId": zod.coerce.number()
})




export const UpdateSubjectBody = zod.object({
  "name": zod.string().min(1).optional(),
  "icon": zod.string().optional(),
  "description": zod.string().optional()
})

export const UpdateSubjectResponse = zod.object({
  "id": zod.number(),
  "name": zod.string(),
  "classId": zod.number(),
  "icon": zod.string().nullish(),
  "description": zod.string().nullish()
})


/**
 * @summary Delete a subject (operator only)
 */
export const DeleteSubjectParams = zod.object({
  "subjectId": zod.coerce.number()
})


/**
 * @summary Create a material (operator only)
 */
export const CreateMaterialParams = zod.object({
  "subjectId": zod.coerce.number()
})





export const CreateMaterialBody = zod.object({
  "title": zod.string().min(1),
  "type": zod.string().min(1),
  "description": zod.string().optional(),
  "url": zod.string().optional(),
  "objectPath": zod.string().optional()
})


/**
 * @summary Update a material (operator only)
 */
export const UpdateMaterialParams = zod.object({
  "materialId": zod.coerce.number()
})




export const UpdateMaterialBody = zod.object({
  "title": zod.string().min(1).optional(),
  "type": zod.string().optional(),
  "description": zod.string().optional(),
  "url": zod.string().optional(),
  "objectPath": zod.string().optional()
})

export const UpdateMaterialResponse = zod.object({
  "id": zod.number(),
  "title": zod.string(),
  "type": zod.string(),
  "subjectId": zod.number(),
  "url": zod.string().nullish(),
  "objectPath": zod.string().nullish(),
  "description": zod.string().nullish()
})


/**
 * @summary Delete a material (operator only)
 */
export const DeleteMaterialParams = zod.object({
  "materialId": zod.coerce.number()
})


/**
 * @summary Request a presigned URL for file upload
 */





export const RequestUploadUrlBody = zod.object({
  "name": zod.string().min(1),
  "size": zod.number().min(1),
  "contentType": zod.string().min(1)
})

export const RequestUploadUrlResponse = zod.object({
  "uploadURL": zod.string(),
  "objectPath": zod.string()
})


/**
 * @summary Serve an uploaded object
 */
export const GetStorageObjectParams = zod.object({
  "objectPath": zod.coerce.string()
})